add native method for sharing data directly

+++ android/app/src/main/kotlin/com/derdilla/blood_pressure_app/StorageProvider.kt
@@ -20,12 +20,16 @@ class StorageProvider(private var context: Context,
          when (call.method) {
              "shareFile" -> shareFile(call.argument<String>("path")!!,
                  call.argument<String>("mimeType")!!, call.argument<String>("name"))
+             "shareData" -> shareData(call.argument<ByteArray>("data")!!,
+                 call.argument<String>("mimeType")!!, call.argument<String>("name")!!)
         }
     }
 
     /**
      * Show the share sheet for saving a file.
      *
+     * @param path The path to the file in internal storage
+     * @param mimeType The mime type that gets send with the intend
      * @param name Used to make the name of the shared file different from the original name.
      */
     private fun shareFile(path: String, mimeType: String, name: String?) {
@@ -41,12 +45,43 @@ class StorageProvider(private var context: Context,
         startActivity(context, shareIntent, null)
     }
 
+    /**
+     * Create a file from data and show start Intent to show a share sheet.
+     *
+     * @param data The data that gets shared
+     * @param mimeType The mime type that gets send with the intend
+     * @param name The name of the shared file
+     */
+    private fun shareData(data: ByteArray, mimeType: String, name: String) {
+        val uri = sharableUriFromData(data, name)
+
+        val sendIntent: Intent = Intent().apply {
+            action = Intent.ACTION_SEND
+            putExtra(Intent.EXTRA_STREAM, uri)
+            type = mimeType
+        }
+
+        val shareIntent = Intent.createChooser(sendIntent, null)
+        startActivity(context, shareIntent, null)
+    }
+
 
     private fun sharableUriFromPath(path: String, name: String?): Uri {
         val initialFile = File(path)
+        if (!initialFile.exists()) throw IllegalArgumentException("Tried to create URI from nonexistent file")
+
         val sharablePath = File(shareFolder, name ?: initialFile.name)
         if (sharablePath.exists()) sharablePath.delete()
         initialFile.copyTo(sharablePath)
+        return getUriForFile(context, "com.derdilla.bloodPressureApp.share", sharablePath)
+    }
+
+    private fun sharableUriFromData(data: ByteArray, name: String): Uri {
+        if (!shareFolder.exists()) shareFolder.mkdirs()
+        val sharablePath = File(shareFolder, name)
+        if (sharablePath.exists()) sharablePath.delete()
+        sharablePath.createNewFile()
+        sharablePath.writeBytes(data)
         return getUriForFile(context, "com.derdilla.bloodPressureApp.share", sharablePath)
     }
 }
